docs: move `overlay` and `progress` examples to `client`
