+++ examples/client/overlay/app.js
