+++ examples/client/overlay/webpack.config.js
@@ -2,9 +2,10 @@
 
 // our setup function adds behind-the-scenes bits to the config that all of our
 // examples need
-const { setup } = require("../util");
+const { setup } = require("../../util");
 
 module.exports = setup({
   context: __dirname,
-  entry: "./app.js",
+  // create error for overlay
+  entry: "./invalid.js",
 });
