+++ examples/client/progress/app.js
