+++ examples/client/progress/webpack.config.js
@@ -2,10 +2,9 @@
 
 // our setup function adds behind-the-scenes bits to the config that all of our
 // examples need
-const { setup } = require("../util");
+const { setup } = require("../../util");
 
 module.exports = setup({
   context: __dirname,
-  // create error for overlay
-  entry: "./invalid.js",
+  entry: "./app.js",
 });
